+++ lib/phoenix/test/channel_test.ex
@@ -13,15 +13,26 @@ defmodule Phoenix.ChannelTest do
   To get started, define the module attribute `@endpoint`
   in your test case pointing to your application endpoint.
 
-  Then you can directly `subscribe_and_join/3` topics and
-  channels:
+  Then you can directly create a socket and
+  `subscribe_and_join/4` topics and channels:
 
       {:ok, _, socket} =
-        subscribe_and_join(RoomChannel, "rooms:lobby", %{"id" => 3})
+        socket("user:id", %{some_assigns: 1})
+        |> subscribe_and_join(RoomChannel, "rooms:lobby", %{"id" => 3})
 
-  The function above will subscribe the current test process
-  to the "rooms:lobby" topic and start a channel in another
-  process. It returns `{:ok, reply, socket}` or `{:error, reply}`.
+  You usually want to set the same ID and assigns your
+  `UserSocket.connect/2` callback would set. Alternatively,
+  you can call your `UserSocket.connect/2` function with an
+  empty socket:
+
+      {:ok, _, socket} =
+        UserSocket.connect(%{"some" => "params"}, socket())
+        |> subscribe_and_join(RoomChannel, "rooms:lobby", %{"id" => 3})
+
+  Once called, `subscribe_and_join/4` will subscribe the
+  current test process to the "rooms:lobby" topic and start a
+  channel in another process. It returns `{:ok, reply, socket}`
+  or `{:error, reply}`.
 
   Now, in the same way the channel has a socket representing
   communication it will push to the client. Our test has a
@@ -132,7 +143,6 @@ defmodule Phoenix.ChannelTest do
   alias Phoenix.Channel.Server
 
   defmodule NoopSerializer do
-
     @behaviour Phoenix.Transports.Serializer
 
     def fastlane!(%Broadcast{} = msg) do
@@ -145,7 +155,6 @@ defmodule Phoenix.ChannelTest do
 
     def encode!(%Reply{} = reply), do: reply
     def encode!(%Message{} = msg), do: msg
-
     def decode!(message, _opts), do: message
   end
 
@@ -157,17 +166,64 @@ defmodule Phoenix.ChannelTest do
   end
 
   @doc """
-  Same as subscribe_and_join/3 but returns either the socket or throws an error
+  Builds a socket.
 
-  This is helpful when you are not testing authentication and just need the
-  socket.
+  The socket is then used to subscribe and join channels.
+  Use this function when you want to create a blank socket
+  to pass to functions like `UserSocket.connect/2`.
+
+  Otherwise, use `socket/2` if you want to build a socket with
+  id and assigns.
+
+  The socket endpoint is read from the `@endpoint` variable.
   """
-  defmacro subscribe_and_join!(channel, topic, payload \\ Macro.escape(%{})) do
-    quote do
-      case subscribe_and_join(@endpoint, unquote(channel), unquote(topic), unquote(payload)) do
-        {:ok, _, socket} -> socket
-        {:error, error} -> raise "Could not join channel. Got error: #{inspect(error)}"
+  defmacro socket() do
+    if endpoint = Module.get_attribute(__CALLER__.module, :endpoint) do
+      quote do
+        %Socket{serializer: NoopSerializer,
+                transport_pid: self(),
+                endpoint: unquote(endpoint),
+                pubsub_server: unquote(endpoint).__pubsub_server__(),
+                transport: unquote(__MODULE__)}
       end
+    else
+      raise "module attribute @endpoint not set for socket/0"
+    end
+  end
+
+  @doc """
+  Builds a socket with given id and assigns.
+
+  The socket endpoint is read from the `@endpoint` variable.
+  """
+  defmacro socket(id, assigns) do
+    if endpoint = Module.get_attribute(__CALLER__.module, :endpoint) do
+      quote do
+        %Socket{serializer: NoopSerializer,
+                transport_pid: self(),
+                endpoint: unquote(endpoint),
+                pubsub_server: unquote(endpoint).__pubsub_server__(),
+                id: unquote(id),
+                assigns: Enum.into(unquote(assigns), %{}),
+                transport: unquote(__MODULE__)}
+      end
+    else
+      raise "module attribute @endpoint not set for socket/2"
+    end
+  end
+
+  @doc """
+  Same as subscribe_and_join/4 but returns either the socket
+  or throws an error.
+
+  This is helpful when you are not testing joining the channel
+  and just need the socket.
+  """
+  def subscribe_and_join!(%Socket{} = socket, channel, topic, payload \\ %{})
+      when is_atom(channel) and is_binary(topic) and is_map(payload) do
+    case subscribe_and_join(socket, channel, topic, payload) do
+      {:ok, _, socket} -> socket
+      {:error, error}  -> raise "could not join channel, got error: #{inspect(error)}"
     end
   end
 
@@ -183,29 +239,11 @@ defmodule Phoenix.ChannelTest do
   linked to the test process.
 
   It returns `{:ok, reply, socket}` or `{:error, reply}`.
-
-  The endpoint is read from the `@endpoint` variable.
   """
-  defmacro subscribe_and_join(channel, topic, payload \\ Macro.escape(%{})) do
-    quote do
-      subscribe_and_join(@endpoint, unquote(channel), unquote(topic), unquote(payload))
-    end
-  end
-
-  @doc """
-  Subscribes to the given topic and joins the channel powered
-  by the pubsub server in endpoint under the given topic and
-  payload.
-
-  This is useful when you need to join a channel in different
-  enpoints, in practice, `subscribe_and_join/3` is recommended.
-  """
-  def subscribe_and_join(endpoint, channel, topic, payload) do
-    unless endpoint do
-      raise "module attribute @endpoint not set for subscribe_and_join/3"
-    end
-    endpoint.subscribe(self(), topic)
-    join(endpoint, channel, topic, payload)
+  def subscribe_and_join(%Socket{} = socket, channel, topic, payload \\ %{})
+      when is_atom(channel) and is_binary(topic) and is_map(payload) do
+    socket.endpoint.subscribe(self(), topic)
+    join(socket, channel, topic, payload)
   end
 
   @doc """
@@ -215,34 +253,10 @@ defmodule Phoenix.ChannelTest do
   which is linked to the test process.
 
   It returns `{:ok, reply, socket}` or `{:error, reply}`.
-
-  The endpoint is read from the `@endpoint` variable.
   """
-  defmacro join(channel, topic, payload \\ Macro.escape(%{})) do
-    quote do
-      join(@endpoint, unquote(channel), unquote(topic), unquote(payload))
-    end
-  end
-
-  @doc """
-  Joins the channel powered by the pubsub server in
-  endpoint under the given topic and payload.
-
-  This is useful when you need to join a channel in
-  different enpoints, in practice, `join/3` is recommended.
-  """
-  def join(endpoint, channel, topic, payload) do
-    unless endpoint do
-      raise "module attribute @endpoint not set for join/3"
-    end
-
-    socket = %Socket{serializer: NoopSerializer,
-                     transport_pid: self(),
-                     endpoint: endpoint,
-                     pubsub_server: endpoint.__pubsub_server__(),
-                     topic: topic,
-                     channel: channel,
-                     transport: __MODULE__}
+  def join(%Socket{} = socket, channel, topic, payload \\ %{})
+      when is_atom(channel) and is_binary(topic) and is_map(payload) do
+    socket = %Socket{socket | topic: topic, channel: channel}
 
     case Server.join(socket, payload) do
       {:ok, reply, pid} ->